Remove running feature and energy attributes
Running is not really that great a feature in this kind of game.
Players want to focus on simply moving around and shooting (mostly),
and not on activating/deactivating the run mode all the time. Keep it
simple.

Maybe at some point some NPC type can have a special behavior that
makes it run really fast for a limited amount of time.