[Debug] Improved path drawing
Draws a circle on the destination tile if the path consists only of a single
tile. Also draws a circle on the creature whose MoveTask it is. The colors
for the circles and lines, and the size of the circle, are chosen randomly
to make it easier to distinguish the paths of different NPCs.